requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
ultralytics==8.3.30
opencv-python-headless==4.10.0.84
python-dotenv==1.0.1
SQLAlchemy==2.0.35
oracledb==2.4.1
google-genai